Backend/requirements.common.txt
Django==2.2.4
django-auth-ldap==2.0.0
django-background-tasks==1.2.0
django-bootstrap4==0.0.8
django-filter==2.2.0
django-tables2==2.1.0
djangorestframework==3.10.2
mysqlclient==1.4.4
numpy==1.17.0
pandas==0.25.0
pylint==2.3.1
Sphinx==2.2.0
SQLAlchemy==1.3.7
coloredlogs==10.0
tablib==0.13.0